+++ apps/docs/src/docs/composables/demo/UseToastSetup.vue
@@ -1,5 +1,0 @@
-<template>
-  <BApp>
-    <router-view />
-  </BApp>
-</template>